refactored log out button
