+++ react-dashboard/src/layouts/home.jsx
@@ -1,4 +1,4 @@
-import { Login } from "../services/login";
+import { Login } from "../services/firebase";
 import "./home.css";
 import background from "../assets/img/sidebar-4.jpg";
